api/requirements.txt
Django==4.2.4
psycopg2-binary==2.9.7
pycparser==2.21
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
djangorestframework
Pillow